docs: add LineChartWithReferenceLines example in TypeScript (.tsx)
